+++ classes/multilingual-plugins/easy-language/class-texts-in-use-table.php
@@ -234,7 +234,11 @@ class Texts_In_Use_Table extends WP_List_Table {
 
 			// show hint for pro in used in column.
 			case 'used_in':
-				return '<span class="pro-marker">' . __( 'Only in Pro', 'easy-language' ) . '</span>';
+				return '<span class="pro-marker">' . __( 'Info only with Pro visible', 'easy-language' ) . '</span>';
+
+			// set default return to empty string.
+			default:
+				return '';
 		}
 	}